fix(ci): line-limit guard — split oversized test part files
